Adding some fancy graphics and text
Been keeping an eye on memory usage. It seems to stablize around the
500MB range. Must be part of the Garbage Collector so it's ignorable
for now.

Right now it cycles between party members. The plan will be that it
shows stats via multiple "pages" so that the info is readable

Page 1: Health, EXP, Level, Typings
Page 2: Attacks and PP
Page 3: Stats (With a EV/IV spread display)

After the last page it will then move to the next Pokemon in the party